+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 ## [Unreleased]
-_No unreleased changes yet._
+- bug: fix bug in `core/add-update-label-weekly.js` that was not returning a timeline when assignee was assigned by others
 
 ## v1.0.3
 2026-02-16